hw3 fix

+++ src/main/java/com/geekbrains/app/controller/ProductController.java
@@ -37,7 +37,7 @@ public class ProductController {
     }
 
     @PostMapping("/create")
-    public String saveProduct(@RequestParam String title, @RequestParam Double cost) {
+    public String saveProduct(@RequestParam String title, @RequestParam Integer cost) {
         Product product = new Product(title, cost);
         productService.save(product);
         return "redirect:/products/show_all";
